separated files and cleaned code up a bit
