+++ budgeting_logic/constants.py
@@ -1,0 +1,28 @@
+COLUMNS = [
+    'id',
+    'date',
+    'description',
+    'amount',
+    'category',
+]
+
+CATEGORIES = {
+        0: 'Restaurants',
+        1: 'Groceries',
+        2: 'Clothing', 
+        3: 'Entertainment', 
+        4: 'Gas', 
+        5: 'Gifts', 
+        6: 'Insurance',
+        7: 'House Stuff', 
+        8: 'Internet', 
+        9: 'Luna Stuff', 
+        10: 'Medical', 
+        11: 'Other', 
+        12: 'Parking', 
+        13: 'Phone', 
+        14: 'Rent/Mortgage', 
+        15: 'Utilities', 
+        16: 'Travel', 
+        17: 'IGNORE', 
+    }
