+++ main.py
@@ -1,53 +1,13 @@
 
 import pandas as pd
 import os
-import gspread
-from google.oauth2.service_account import Credentials
-from accounts.american_express import AmericanExpressRow
+
 from accounts.account_utils import get_date_string
-from colorama import Fore, Style
+from budgeting_logic.logic import categorize_rows, get_cleaned_df
+from budgeting_logic.constants import COLUMNS
+from budgeting_logic.google_sheets import get_worksheet, get_existing_transactions
 
-COLUMNS = [
-    'id',
-    'date',
-    'description',
-    'amount',
-    'category',
-]
-
-CATEGORIES = {
-        0: 'Restaurants',
-        1: 'Groceries',
-        2: 'Clothing', 
-        3: 'Entertainment', 
-        4: 'Gas', 
-        5: 'Gifts', 
-        6: 'Insurance',
-        7: 'House Stuff', 
-        8: 'Internet', 
-        9: 'Luna Stuff', 
-        10: 'Medical', 
-        11: 'Other', 
-        12: 'Parking', 
-        13: 'Phone', 
-        14: 'Rent/Mortgage', 
-        15: 'Utilities', 
-        16: 'Travel', 
-        17: 'IGNORE', 
-    }
-
-
-def get_cleaned_df(df: pd.DataFrame): 
-     # Step 2b: Clean the data and populate the new dataframe.
-    cleaned_rows = []
-    for index, row in df.iterrows(): 
-        amex_row = AmericanExpressRow().clean_data(row)
-        cleaned_rows.append(amex_row.convert_to_list())
-
-    return pd.DataFrame(cleaned_rows, columns=COLUMNS)
-
-
-def process(df: pd.DataFrame, existing_sheet_df: pd.DataFrame): 
+def process_file(df: pd.DataFrame, existing_sheet_df: pd.DataFrame): 
     # Step 2: Clean data 
     # TODO(ncarchio): Step 2a: figure out what class to call (use the headers to figure it out 
     # -- each bank/credit card has its own headers)
@@ -72,7 +32,7 @@ def process(df: pd.DataFrame, existing_sheet_df: pd.DataFrame):
     print(f"deduped: {df_deduped}, len: {len(df_deduped)}")
 
 
-    # Now, get all new entries after deduping. 
+    # Get all new entries after deduping. 
     number_of_deduped_entries = (len(new_df) + db_entries_length) - len(df_deduped)
     number_of_new_entries = len(new_df) - number_of_deduped_entries
     start_index = len(df_deduped) - number_of_new_entries
@@ -86,7 +46,7 @@ def process(df: pd.DataFrame, existing_sheet_df: pd.DataFrame):
     new_entries_df['date'] = new_entries_df['date'].apply(lambda date: get_date_string(date))
 
     # Categorize all rows if they don't already have a category. 
-    # 1. Try to 'auto' categorize based off of the Categories map
+    # TODO(ncarchio): 1. Try to 'auto' categorize based off of the Categories map
     #   - within this, we should be able to automatically categorize and if we find it is 
     #    an 'IGNORE', we remove that from the dataframe
     # 2. If we can't auto categorize, ask user for manual input. 
@@ -97,54 +57,24 @@ def process(df: pd.DataFrame, existing_sheet_df: pd.DataFrame):
     worksheet = get_worksheet()
     worksheet.append_rows(finalized_new_transactions)
 
-    
-def get_worksheet(): 
-    sheets = workbook.worksheets()
-    print(sheets)
-    return workbook.worksheet('Transactions')
+    print("""
+    -----------------------------------------------------------------------
+    Thanks for using bougiet, now, head over to Google Sheets and enjoy :) 
+    -----------------------------------------------------------------------
+    """)
 
-
-def get_existing_transactions(): 
-    raw_transactions = get_worksheet().get_all_values()
-    return pd.DataFrame(raw_transactions[1:], columns=COLUMNS)
-
-
-def get_category_input(description: str): 
-    print(Fore.YELLOW + f"It's time to categorize all of your transactions. Please add your input below." + Style.RESET_ALL)
-    print(Fore.GREEN + f"Enter the number of one of the following categories.\nCurrent transaction: {description}" + Style.RESET_ALL)
-    for key, value in CATEGORIES.items(): 
-        print(f"{key}: {value}")
-    category_input = int(input(Fore.CYAN + f"\nCategory: " + Style.RESET_ALL))
-    return category_input
-
-
-def categorize_rows(df: pd.DataFrame):
-    for index, row in df.iterrows(): 
-        # Skip already categorized rows.
-        if len(df.loc[index, 'category']) > 0: 
-            continue
-        category_input = get_category_input(row.description)
-        df.loc[index,'category'] = CATEGORIES[category_input]
-
-    print(f"df after: {df}")
-    return df
 
 
 def main(): 
-    # Step 1: Input the file.
-    # Get the directory of the current script
+    # Step 1: Fetch the input files.
     current_dir = os.path.dirname(os.path.abspath(__file__))
-
-    # Get the folder path for the CSVs
     test_csv_folder_path = os.path.join(current_dir, "test_csvs/") 
-
-    # Get a list of all CSV files in the folder
     csv_files = [f for f in os.listdir(test_csv_folder_path) if f.endswith('.csv')]
 
-    # Get all the transactions from sheets. 
+    # Step 2: Get all the transactions from sheets. 
     existing_sheet_df = get_existing_transactions()
     
-    # Loop through each CSV file
+    # Step 3: Loop through each CSV file.
     for file in csv_files:
         file_path = os.path.join(test_csv_folder_path, file)
         print(f"processing {file_path}")
@@ -153,12 +83,9 @@ def main():
         print(f"existing df: {existing_sheet_df}")
         print(f"new df: {df}")
 
-        process(df, existing_sheet_df)
+        process_file(df, existing_sheet_df)
         
-        
-
-
-# run main
+# Run the main function
 main()
